Should we implement Inherited widget for FAcoordionItem
