+++ forui/lib/widgets/accordion.dart
@@ -1,8 +1,10 @@
 /// {@category Widgets}
 ///
-/// An interactive heading that reveals a section of content.
+/// A vertically stacked set of interactive headings that each reveal a section of content.
 ///
 /// See https://forui.dev/docs/accordion for working examples.
 library forui.widgets.accordion;
 
-export '../src/widgets/accordion.dart';
+export '../src/widgets/accordion/accordion.dart';
+export '../src/widgets/accordion/accordion_controller.dart';
+export '../src/widgets/accordion/accordion_item.dart';
